// Project/InternalAuditSystem.Web/ClientApp/src/app/swaggerapi/AngularFiles/model/entityRelationMappingAC.ts
/**
 * My Title
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AuditableEntityAC } from './auditableEntityAC';
import { BaseModelAC } from './baseModelAC';
import { RelationshipTypeAC } from './relationshipTypeAC';

export interface EntityRelationMappingAC extends BaseModelAC { 
    entityId: string;
    auditableEntity?: string;
    relationTypeId: string;
    relatedEntityName?: string;
    relationName?: string;
    relatedEntityId: string;
    auditableEntityACList?: Array<AuditableEntityAC>;
    relationshipTypeACList?: Array<RelationshipTypeAC>;
}